MSVC projects: explicitly define WINVER, _WIN32_WINNT and NTDDI_VERSION to 0x0501 and 0x0502 for 32 and 64 bits targets respectively.
This ensures that functions incompatible with Windows XP cannot be accidentally used because they will be #ifdef'd out in the Windows headers. The exception is ZydisWinKernel, which requires Windows 7 mostly due to WDK toolchain limitations.

Note that the projects still use the v141 toolset and not v141_xp as might be expected. The latter toolset is a giant hack that requires additional XP targeting support to be installed with Visual Studio, and changes the platform SDK to the ancient Windows 7 SDK as well as adding some other undesirable build hacks. In practice however the current toolchain outputs binaries that work perfectly fine on platforms from XP through 10, so long as the above preprocessor defines are correctly set to the minimum desired target version