Remove ZydisPE
The PE parsing code is wildly unsafe and doesn't perform any input validation.
It was originally written as an example on how to use our formatter API for
using custom symbols in disassembly output. However, living in the `tools`
directory incorrectly suggested to users that it's a safe tool intended for
actual public use.

Following a security report, we decided to just delete it for the time being,
possibly bringing it back later.

We intentionally never defined install rules for it (it's not installed on
`make install`) and also don't ship it in the packages of any package manager
repository, so the practical security impact should be low.