docs: add man pages
Zydis ships two command line tools, but doesn't provide any man page.
Since Debian enforces that every binary must also provide a man page
here I am providing a patch.

Standard UNIX manpages are painful to write, but luckly there are tools
like Ronn(-NG) or Pandoc that allow to write the page in simple Markdown
and convert it in roff, the format that `man` uses.

These man pages are extremely basic, since I'm not very familiar with
the tools and how they work :/
Please don't hesitate making suggestions :)

Since Ronn-NG is an extra dependency the manpages are not built by
default, but only if requested with `ZYDIS_BUILD_MAN`.

Ronn-NG is required instead of the old Ronn because the latter always
generates output files in the same dir as the input files, and that
doesn't play nice with CMake and out of tree builds, while the -NG
version has a `--output-dir` option.