Cache For Common Instructions
The use of the common instruction cache yields an average speed increase of 15% to 20% on my computer. It does, however, slightly reduce the speed if it is not used as it has to check if the cache is being used, so perhaps this could be either enabled/disabled through a #define or have it be enabled by default. It could probably be further optimized for even greater speeds. It is very simple to enable for a single instance of a ZydisDecoder, and I have included commented out calls to the initialization in the Formatter examples.