CI: switch to an include-only way of defining the build matrix
Our previous configuration was built on incorrect assumptions about how
`include` works in GitHub actions. Various builds that we wanted to run
were not actually executed, and multiple additional ones that we didn't
want were.

For our purposes it's a lot simpler to reason about when just writing out
the full list of tests to run manually.