Update MSVC projects to VS 2022
NB: The ZydisWinKernel sample still compiles for, and runs on Windows 7 x64 and higher as before. However, some adjustments were made in order for this to work, because the VS2022 WDK officially only recognises 'Windows 10 and higher' as valid target OS. 32 bit support for this sample has NOT survived this update due to new WDK restrictions.