Fix -Wtype-limits warnings by removing redundant assertions
This was uncovered in the context of WebKit. These assertions
are always true as the bitfield variables in question are
too small to hold the values that the assertions are comparing
against:

- ZydisSetAttributes: The arrays are defined to be larger than
the bitfield holding cpu_state, fpu_state, and xmm_state,
so the comparison is always true.

- ZydisSetEffectiveOperandWidth: operand_size_map is a 3-bit bitfield,
and the arrays have a size 8, thus the comparison is always true.

- ZydisGetOperandDefinitions: operand_reference is a 15-bit bitfield,
it can't hold 0xFFFF (which needs 16 bits), so the comparison is always
true.

See also https://bugs.webkit.org/show_bug.cgi?id=252309