Doc: extract documentation from all headers
We previously went with explicitly adding `@file` directives to our
headers instead of using `EXTRACT_ALL` because the latter was broken
with the m.css theme that we were using previously. m.css has a custom
code generator that didn't support the directive. We've since switched
to another theme that just re-styles the vanilla HTML, so we're fine
to just use that now (instead of changing the code-gen to add the
directives in all generated headers).